git-gui: Slightly tweak new window geometry.
I didn't really like the way a new git-gui launched in a new repository
as the window geometry wasn't quite the best layou.  So this is a minor
tweak to try and get space distributed around the window better.

By decreasing the widths we're also able to shrink the gui smaller
without Tk clipping content at the edge of the window.  A nice feature.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>