git-gui: Implement basic branch switching through read-tree.
If the user selects a different branch from the Branch menu, or asks
us to create a new branch and immediately checkout that branch we
now perform the update of the working directory by way of a 2 way
read-tree invocation.

This emulates the behavior of `git checkout branch` or the behavior
of `git checkout -b branch initrev`.  We don't however support the
-m style behavior, where a switch can occur with file level merging
performed by merge-recursive.  Support for this is planned for a
future update.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>