git-gui: Force focus to the diff viewer on mouse click.
Apparently a "feature" of Tcl/Tk on Mac OS X is that a disabled text
widget cannot receive focus or receive a selection within it.  This
makes the diff viewer almost useless on that platform as you cannot
select individual parts of the buffer.

Now we force focus into the diff viewer when its clicked on with
button 1.  This works around the feature and allows selection to
work within the viewer just like it does on other less sane systems,
like Microsoft Windows.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>